+++ asphalt/mailer/component.py
@@ -1,34 +1,38 @@
 import logging
 from typing import Dict, Any
 
+from asphalt.core import Component, Context, PluginContainer, merge_config, context_teardown
+from async_generator import yield_
 from typeguard import check_argument_types
 
-from asphalt.core.component import Component
-from asphalt.core.context import Context
-from asphalt.core.util import PluginContainer, merge_config
 from asphalt.mailer.api import Mailer
 
 mailer_backends = PluginContainer('asphalt.mailer.mailers', Mailer)
-
 logger = logging.getLogger(__name__)
 
 
 class MailerComponent(Component):
     """
-    Publishes one or more :class:`~asphalt.mailer.api.Mailer` resources.
+    Creates one or more :class:`~asphalt.mailer.api.Mailer` resources.
 
-    If more than one mailer is to be configured, provide a ``mailers`` argument as a dictionary
-    where the key is the resource name and the value is a dictionary of keyword arguments to
-    :meth:`configure_mailer`. Otherwise, directly pass those keyword arguments to the component
-    constructor itself.
+    Mailers can be configured in two ways:
 
-    If ``mailers`` is defined, any extra keyword arguments are used as default values for
-    :meth:`configure_mailer` for all mailers (:func:`~asphalt.core.util.merge_config` is used to
-    merge the per-mailer arguments with the defaults). Otherwise, a single mailer is created based
-    on the provided default arguments, with ``context_attr`` defaulting to ``mailer``.
+    #. a single mailer, with configuration supplied directly as keyword arguments to this
+        component's constructor (with the resource name being ``default`` and the context attribute
+        matching the backend name)
+    #. multiple mailers, by providing the ``mailers`` option where each key is the resource
+        name and each value is a dictionary containing that mailer's configuration (with the
+        context attribute matching the resource name by default)
 
-    :param mailers: a dictionary of resource name ⭢ :meth:`configure_mailer` keyword arguments
-    :param default_mailer_args: default values for :meth:`configure_mailer`
+    Each mailer configuration has two special options that are not passed to the constructor of
+    the backend class:
+
+    * backend: entry point name of the mailer backend class (required)
+    * context_attr: name of the context attribute of the mailer resource
+
+    :param mailers: a dictionary of resource name ⭢ constructor arguments for the chosen
+        backend class
+    :param default_mailer_args: default values for constructor keyword arguments
     """
 
     def __init__(self, mailers: Dict[str, Dict[str, Any]] = None, **default_mailer_args):
@@ -41,28 +45,20 @@ class MailerComponent(Component):
         self.mailers = []
         for resource_name, config in mailers.items():
             config = merge_config(default_mailer_args, config)
-            config.setdefault('context_attr', resource_name)
-            context_attr, mailer = self.configure_mailer(**config)
+            backend = config.pop('backend')
+            context_attr = config.pop('context_attr', resource_name)
+            mailer = mailer_backends.create_object(backend, **config)
             self.mailers.append((resource_name, context_attr, mailer))
 
-    @classmethod
-    def configure_mailer(cls, context_attr: str, backend: str, **backend_kwargs):
-        """
-        Configure a Mailer backend with the given parameters.
-
-        :param context_attr: the mailer's attribute name on the context (defaults to the resource
-            name)
-        :param backend: specifies the type of mailer (declared as entry points in the
-            ``asphalt.mailer.mailers`` namespace)
-        :param backend_kwargs: keyword arguments passed to the constructor of the backend class
-
-        """
-        assert check_argument_types()
-        mailer = mailer_backends.create_object(backend, **backend_kwargs)
-        return context_attr, mailer
-
+    @context_teardown
     async def start(self, ctx: Context):
         for resource_name, context_attr, mailer in self.mailers:
             await mailer.start(ctx)
-            ctx.publish_resource(mailer, resource_name, context_attr, types=[Mailer])
-            logger.info('Configured mailer (%s / ctx.%s)', resource_name, context_attr)
+            ctx.add_resource(mailer, resource_name, context_attr, [Mailer, type(mailer)])
+            logger.info('Configured mailer (%s / ctx.%s; class=%s)', resource_name, context_attr,
+                        mailer.__class__.__name__)
+
+        await yield_()
+
+        for resource_name, context_attr, mailer in self.mailers:
+            logger.info('Mailer (%s) stopped', resource_name)