+++ tests/test_component.py
@@ -9,14 +9,16 @@ from asphalt.mailer.component import MailerComponent
 @pytest.mark.asyncio
 async def test_component_single(caplog, backend):
     component = MailerComponent(backend=backend)
-    context = Context()
-    await component.start(context)
+    async with Context() as ctx:
+        await component.start(ctx)
+        assert isinstance(ctx.mailer, Mailer)
 
-    assert isinstance(context.mailer, Mailer)
     records = [record for record in caplog.records if record.name == 'asphalt.mailer.component']
     records.sort(key=lambda r: r.message)
-    assert len(records) == 1
-    assert records[0].message.startswith("Configured mailer (default / ctx.mailer)")
+    assert len(records) == 2
+    assert records[0].message.startswith(
+        "Configured mailer (default / ctx.mailer; class=%s)" % ctx.mailer.__class__.__name__)
+    assert records[1].message.startswith('Mailer (default) stopped')
 
 
 @pytest.mark.asyncio
@@ -25,13 +27,17 @@ async def test_component_multiple(caplog):
         'smtp': {'backend': 'smtp', 'context_attr': 'mailer1'},
         'sendmail': {'backend': 'sendmail', 'context_attr': 'mailer2'}
     })
-    context = Context()
-    await component.start(context)
+    async with Context() as ctx:
+        await component.start(ctx)
+        assert isinstance(ctx.mailer1, Mailer)
+        assert isinstance(ctx.mailer2, Mailer)
 
-    assert isinstance(context.mailer1, Mailer)
-    assert isinstance(context.mailer2, Mailer)
     records = [record for record in caplog.records if record.name == 'asphalt.mailer.component']
     records.sort(key=lambda r: r.message)
-    assert len(records) == 2
-    assert records[0].message.startswith("Configured mailer (sendmail / ctx.mailer2)")
-    assert records[1].message.startswith("Configured mailer (smtp / ctx.mailer1)")
+    assert len(records) == 4
+    assert records[0].message.startswith(
+        "Configured mailer (sendmail / ctx.mailer2; class=SendmailMailer)")
+    assert records[1].message.startswith(
+        "Configured mailer (smtp / ctx.mailer1; class=SMTPMailer)")
+    assert records[2].message.startswith('Mailer (sendmail) stopped')
+    assert records[3].message.startswith('Mailer (smtp) stopped')