+++ tests/test_utils.py
@@ -1,6 +1,6 @@
 from email.message import EmailMessage
 
-from asphalt.mailer.util import get_recipients
+from asphalt.mailer.utils import get_recipients
 
 
 def test_get_recipients():